# src/main/java/com/daroch/booking/dto/ErrorResponse.java
package com.daroch.booking.dto;

import java.time.Instant;
import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Data;
import lombok.NoArgsConstructor;

@Data
@Builder
@AllArgsConstructor
@NoArgsConstructor
public class ErrorResponse {

  private Instant timestamp;
  private int status;
  private String errorCode;
  private String message;
  private String path;
  private String traceId;
}